move onceAuthenticated into authenticateViaTwitter
