+++ src/tests/integration/already-authenticated.test.ts
@@ -1,12 +1,10 @@
 import { createMocks } from './mocks'
 import { runBackground } from './steps/run-background'
 import { mountPopup } from './steps/mount-popup'
-import { onceAuthenticated } from './steps/once-authenticated'
 
 describe('already authenticated when extension loads', () => {
   const mocks = createMocks()
 
   runBackground(mocks, { alreadyAuthenticated: true })
   mountPopup(mocks, { alreadyAuthenticated: true, handle: 'exists' })
-  onceAuthenticated(mocks)
 })
