+++ src/tests/integration/happy-path.test.ts
@@ -2,7 +2,6 @@ import { createMocks } from './mocks'
 import { runBackground } from './steps/run-background'
 import { mountPopup } from './steps/mount-popup'
 import { authenticateViaTwitter } from './steps/authenticate-via-twitter'
-import { onceAuthenticated } from './steps/once-authenticated'
 import { signInViaTwitter } from './steps/sign-in-via-twitter'
 import { captureFirstScreenshot } from './steps/capture-first-screenshot'
 import { takingScreenshots } from './steps/taking-screenshots'
@@ -16,7 +15,6 @@ describe('happy path', () => {
   mountPopup(mocks, { handle: 'exists' })
   signInViaTwitter(mocks)
   authenticateViaTwitter(mocks)
-  onceAuthenticated(mocks)
   captureFirstScreenshot(mocks)
   takingScreenshots(mocks)
   feedbackEditing(mocks, { handle: '@zing' })
