+++ src/tests/integration/post-feedback-500.test.ts
@@ -2,7 +2,6 @@ import { createMocks } from './mocks'
 import { runBackground } from './steps/run-background'
 import { mountPopup } from './steps/mount-popup'
 import { authenticateViaTwitter } from './steps/authenticate-via-twitter'
-import { onceAuthenticated } from './steps/once-authenticated'
 import { signInViaTwitter } from './steps/sign-in-via-twitter'
 import { captureFirstScreenshot } from './steps/capture-first-screenshot'
 import { postingFeedback } from './steps/posting-feedback'
@@ -15,7 +14,6 @@ describe('post feedback 500', () => {
   mountPopup(mocks, { handle: 'exists' })
   signInViaTwitter(mocks)
   authenticateViaTwitter(mocks)
-  onceAuthenticated(mocks)
   captureFirstScreenshot(mocks)
   feedbackEditing(mocks, { handle: '@zing' })
   postingFeedback(mocks, { handle: '@zing', result: '500' })
