+++ src/tests/integration/unauthorized.test.ts
@@ -2,7 +2,6 @@ import { createMocks } from './mocks'
 import { runBackground } from './steps/run-background'
 import { mountPopup } from './steps/mount-popup'
 import { authenticateViaTwitter } from './steps/authenticate-via-twitter'
-import { onceAuthenticated } from './steps/once-authenticated'
 import { signInViaTwitter } from './steps/sign-in-via-twitter'
 import { captureFirstScreenshot } from './steps/capture-first-screenshot'
 import { postingFeedback } from './steps/posting-feedback'
@@ -15,7 +14,6 @@ describe('unauthorized', () => {
   mountPopup(mocks, { handle: 'exists' })
   signInViaTwitter(mocks)
   authenticateViaTwitter(mocks)
-  onceAuthenticated(mocks)
   captureFirstScreenshot(mocks)
   feedbackEditing(mocks, { handle: '@zing' })
   postingFeedback(mocks, { handle: '@zing', result: 'unauthorized' })
